requirements.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
python-dotenv
google-generativeai
PyPDF2
sentence-transformers 
PyMuPDF
jinja2
aiofiles